+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Information sections to documentation about importing `snforge_std`
 - Print support for basic numeric data types
 - Functions `parse_txt` and `TxtParser<T>::deserialize_txt` to load data from plain text files and serialize it
+- `get_class_hash` cheatcode
 
 #### Changed